Move CurrentCompanyDefault to companies/api/serializers/defaults.py.
